external/opal-prd: Implement PNOR callbacks
The PNOR read and write calls are implemented by accessing the PNOR
through a mtd device.

The mtd device exposes the entire pnor, however the callback operates on
named partitions. Therefore we use libffs from libflash in the skiboot
tree to parse the partitions and find where to read and write.

Signed-off-by: Joel Stanley <joel@jms.id.au>

Conflicts:
	external/opal-prd/Makefile
	external/opal-prd/opal-prd.c