hw/fake-rtc: Add support for emulated rtc clock
This patch adds support for an emulated rtc clock for generic
platform (e.g BML) which doesn't have support for rtc. Current
implementation provides a fake rtc clock for such platforms which is
driven by device tree provided pointer(ibm,fake-rtc) to bcd coded
date-time values. In absence of this rtc support is not provided.

The patch overcomes this limitation by emulating an rtc clock that's
driven by a periodically scheduled timer. If the device tree is missing
node "ibm,fake-rtc" it creates the periodic timer and on each tick the
rtc counter is incremented.

Signed-off-by: Vaibhav Jain <vaibhav@linux.vnet.ibm.com>